Change NoWLBroadcastSubset to float64

+++ experiments.go
@@ -9,5 +9,5 @@ type Experiments struct {
 	P2pHttpProxy         bool
 	StrategicProviding   bool
 	DisableWLBroadcast   bool
-	NoWLBroadcastSubset  uint
+	NoWLBroadcastSubset  float64
 }
